Change float to decimal in all columns

+++ fin-life-api/database/migrations/1633814143504-create_assets_table.ts
@@ -37,7 +37,7 @@ export class createAssetsTable1633814143504 implements MigrationInterface {
           },
           {
             name: 'all_time_high_price',
-            type: 'float'
+            type: 'decimal'
           },
           {
             name: 'currency',
