+++ fin-life-api/database/migrations/1747044980802-create_market_index_historical_data_table.ts
@@ -31,7 +31,7 @@ export class CreateMarketIndexHistoricalDataTable1747044980802 implements Migrat
           },
           {
             name: 'value',
-            type: 'float'
+            type: 'decimal'
           }
         ],
         indices: [
